feat(reporter-publish): Use Reporter.publish instead of .reporter
Previous implementation used the '.reporter(file, errors)' function on 'coffeelint-stylish' to render results, however this is a non-standard method for CoffeeLint reporters. Only '.ctor(ErrorReport, opts)' and '.publish()' are standard functions.

Enhanced the implementation to use CoffeeLint's ErrorReport feature and harness the '.publish' functionality within 'coffeelint-stylish'.

This is a required change to support external or CoffeeLint's built-in reporters.